backend/requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
lxml
google-genai
pydantic
pydantic-settings
python-dotenv
